refactor post restore code
